chore: rename util/ to common/
